Project reorganized. Installation instructions added to CMakeList file
